Exception handling

+++ src/main/java/com/devsuperior/desafio_crud_client/services/ClientService.java
@@ -3,15 +3,17 @@ package com.devsuperior.desafio_crud_client.services;
 import com.devsuperior.desafio_crud_client.dto.ClientDTO;
 import com.devsuperior.desafio_crud_client.entities.Client;
 import com.devsuperior.desafio_crud_client.repositories.ClientRepository;
+import com.devsuperior.desafio_crud_client.services.exceptions.DatabaseException;
 import com.devsuperior.desafio_crud_client.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ClientService {
@@ -43,15 +45,31 @@ public class ClientService {
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
 
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
-    }
+        try{
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
 
-    @Transactional
-    public void delete(Long id){
-      repository.deleteById(id);
+        }
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id))
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        try {
+            repository.deleteById(id);
+        }
+        catch (EmptyResultDataAccessException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
